BamTools: use longs for all metrics

+++ bam-tools/src/main/java/com/hartwig/hmftools/bamtools/metrics/FlagStats.java
@@ -18,7 +18,7 @@ public class FlagStats
     }
 
     public List<FlagQCStats> typeStats() { return mTypeStats; }
-    public int passCount(final FlagStatType type) { return mTypeStats.get(type.ordinal()).getPassed(); }
+    public long passCount(final FlagStatType type) { return mTypeStats.get(type.ordinal()).getPassed(); }
 
     public void merge(final FlagStats other)
     {
